rename config.py to test_data.py
