+++ pages/books_page.py
@@ -2,7 +2,6 @@ __author__ = 'Alexandra Dunga'
 
 from selenium.webdriver.common.by import By
 
-from config.config import TestData
 from pages.base_page import BasePage
 from pages.books_details_page import BooksDetailsPage
