+++ tests/books_page/test_books_page.py
@@ -2,9 +2,7 @@ __author__ = 'Alexandra Dunga'
 
 import pytest
 
-from config.config import TestData
-from pages.base_page import BasePage
-from pages.books_details_page import BooksDetailsPage
+from config.test_data import TestData
 from pages.books_page import BooksPage
 from tests.test_base import BaseTest
