+++ tests/test_base.py
@@ -2,7 +2,7 @@ __author__ = 'Alexandra Dunga'
 
 import pytest
 
-from config.config import TestData
+from config.test_data import TestData
 from regions.menu_region import MenuRegion
